taxa/suggest endpoint, refactor to avoid circular depencency.
The shared modules are meant to avoid two modules requiring each other, b/c I
was adding code to TaxaController that was requiring ObservationsController,
which already requires TaxaController, and I inceptioned myself into a seizure.
The downside is this sort of required some backcountry surgery, and
ObservationsController, TaxaController, and the Observation model got gutted.
I'm not sure if this is an improvement, but it does allow the kind of code
sharing I needed without more callbacks getting passed on forever.